tests: Break out script test environment
Extract the fake environment used in atvremote tests into its own
module, so it can be re-used by other scripts.
